Prevent search errors

+++ src/components/navbar/SearchModal.tsx
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect, useState } from 'react';
-import { Button, Checkbox, Form, Modal } from 'semantic-ui-react';
+import { Button, Checkbox, Form, Modal, Popup } from 'semantic-ui-react';
 import { SearchFilters } from '../../models/SearchFilters.model';
 import { AddressSearch } from './AddressSearch';
 import { AddressSearchContext } from '../../contexts/AddressSearchContext';
@@ -22,23 +22,24 @@ export const SearchModal = () => {
     let tempFilters = tempSearchFilters;
     tempFilters.filterWord = value;
     setTempSearchFilters(tempFilters);
-    console.log(value)
   };
 
   const handleSubmit = () => {
-    setOpen(false);
-    setSearchFilters(tempSearchFilters);
+    if (tempSearchFilters.address.street) {
+      setOpen(false);
+      setSearchFilters(tempSearchFilters);
 
-    // Save address and geolocation in localStorage
-    if (saveAddress) {
-      const address = {
-        street: tempSearchFilters.address.street,
-        location: tempSearchFilters.address.location,
-      };
-      const localStorage = window.localStorage;
-      localStorage.setItem('address', JSON.stringify(address));
-    } else {
-      localStorage.clear();
+      // Save address and geolocation in localStorage
+      if (saveAddress) {
+        const address = {
+          street: tempSearchFilters.address.street,
+          location: tempSearchFilters.address.location,
+        };
+        const localStorage = window.localStorage;
+        localStorage.setItem('address', JSON.stringify(address));
+      } else {
+        localStorage.clear();
+      }
     }
   };
 
@@ -54,12 +55,16 @@ export const SearchModal = () => {
       onClose={() => setOpen(false)}
       onOpen={() => setOpen(true)}
       open={open}
-      trigger={<Button style={{marginLeft: '-36px'}}>Search</Button>}
+      trigger={<Button style={{ marginLeft: '-36px' }}>Search</Button>}
     >
       <Modal.Header>Search for Restaurants</Modal.Header>
       <Modal.Content>
         <Form onSubmit={handleSubmit}>
-          <Form.Input label='Address' required onChange={handleFilterWordChange}>
+          <Form.Input
+            label='Address'
+            required
+            onChange={handleFilterWordChange}
+          >
             <AddressSearchContext.Provider
               value={[tempSearchFilters, setTempSearchFilters]}
             >
@@ -86,14 +91,22 @@ export const SearchModal = () => {
       </Modal.Content>
       <Modal.Actions>
         <Button onClick={() => setOpen(false)}>Cancel</Button>
-        <Button
-          onClick={() => {
-            handleSubmit();
-          }}
-          positive
-        >
-          Start Search
-        </Button>
+        <Popup
+          trigger={
+            <Button
+              onClick={() => {
+                handleSubmit();
+              }}
+              positive
+            >
+              Start Search
+            </Button>
+          }
+          disabled={tempSearchFilters.address.street !== null}
+          content={'Address is required'}
+          on='click'
+          position='top right'
+        />
       </Modal.Actions>
     </Modal>
   );
